revert directory change
